Proper response are now sent
There was a bug where only the base level response will be sent on the login method. This was due to the fact the the method was returning the value before the SQL query was one updating. To fix this I used a callback methods, created a response method within index.js, and called the response method inside the return value for that call back method.